remove: #333 - DetailView, Store 제거
