--- Spoony-iOS/Spoony-iOS/Source/Feature/Scene/Detail/Presentation/DetailView.swift
@@ -1,385 +1,0 @@
-//
-//  DetailView.swift
-//  SpoonMe
-//
-//  Created by 이명진 on 1/2/25.
-//
-
-import SwiftUI
-
-//TODO: PostView 로 이전 끝나면 제가 삭제할게요!!
-struct DetailView: View {
-    
-    // MARK: - Properties
-    
-    @EnvironmentObject private var navigationManager: NavigationManager
-    
-    // ObservableObject 쓰면 안돼! 초기화가 되버림.
-    // StateObeject 는 초기화가 안됌 상태가 유지가 됌
-    @StateObject private var store: DetailViewStore = DetailViewStore()
-    
-    let postId: Int
-    
-    init(postId: Int) {
-        self.postId = postId
-    }
-    
-    private let userImage = Image(.icGpsMain)
-    
-    @State private var isPresented: Bool = false
-    @State private var popUpIsPresented: Bool = false
-    @State private var toastMessage: Toast?
-    
-    // MARK: - body
-    
-    var body: some View {
-        ZStack {
-            VStack(spacing: 0) {
-                CustomNavigationBar(
-                    style: .detailWithChip,
-                    spoonCount: store.state.spoonCount,
-                    onBackTapped: {
-                        navigationManager.pop(1)
-                    }
-                )
-                ScrollView(.vertical) {
-                    VStack(spacing: 0) {
-                        userProfileSection
-                        imageSection
-                        reviewSection
-                        placeInfoSection
-                    }
-                }
-                .simultaneousGesture(dragGesture)
-                .onTapGesture {
-                    dismissDropDown()
-                }
-                .overlay(alignment: .topTrailing, content: {
-                    dropDownView
-                })
-                .scrollIndicators(.hidden)
-                .toastView(toast: $toastMessage)
-                .onAppear {
-                    store.send(intent: .fetchInitialValue(postId: postId))
-                    
-                    if !store.state.successService {
-                        navigationManager.pop(1)
-                    }
-                    
-                }
-                .onChange(of: store.state.toast) { _, newValue in
-                    toastMessage = newValue
-                }
-                
-                bottomView
-                    .frame(height: 80.adjustedH)
-            }
-            .toolbar(.hidden, for: .tabBar)
-            
-            if store.state.isLoading {
-                ZStack {
-                    ProgressView()
-                        .progressViewStyle(CircularProgressViewStyle())
-                }
-                .transition(.opacity)
-                .animation(.easeInOut, value: store.state.isLoading)
-            }
-        }
-    }
-}
-
-// MARK: Gesture
-extension DetailView {
-    private var dragGesture: some Gesture {
-        DragGesture()
-            .onChanged { _ in
-                dismissDropDown()
-            }
-    }
-}
-
-// MARK: Method
-extension DetailView {
-    private func dismissDropDown() {
-        isPresented = false
-    }
-}
-
-// MARK: - Subviews
-
-extension DetailView {
-    private var userProfileSection: some View {
-        HStack(alignment: .center, spacing: 14.adjustedH) {
-            
-            RemoteImageView(urlString: store.entity.userImageUrl)
-                .scaledToFit()
-                .clipShape(Circle())
-                .frame(width: 48.adjusted, height: 48.adjustedH)
-            
-            VStack(alignment: .leading, spacing: 4.adjustedH) {
-                Text(store.entity.userName)
-                    .customFont(.body2b)
-                    .foregroundStyle(.black)
-                
-                Text(store.entity.regionName)
-                    .customFont(.caption1m)
-                    .foregroundStyle(.gray400)
-            }
-            
-            Spacer()
-            
-            if !store.entity.isMine {
-                Image(.icMenu)
-                    .onTapGesture {
-                        isPresented.toggle()
-                    }
-            }
-            
-        }
-        .padding(EdgeInsets(top: 8.adjustedH, leading: 20.adjusted, bottom: 8.adjustedH, trailing: 20.adjusted))
-        .padding(.bottom, 24.adjustedH)
-    }
-    
-    @ViewBuilder
-    private var imageSection: some View {
-        let imageList: [String] = store.entity.photoUrlList
-        
-        if imageList.isEmpty {
-            Rectangle()
-                .foregroundStyle(.gray400)
-                .frame(width: 335.adjusted, height: 335.adjustedH)
-                .cornerRadius(11.16)
-                .padding(EdgeInsets(top: 0, leading: 20.adjusted, bottom: 32.adjustedH, trailing: 20.adjusted))
-        } else if imageList.count == 1 {
-            RemoteImageView(urlString: imageList[0])
-                .scaledToFill()
-                .frame(width: 335.adjusted, height: 335.adjustedH)
-                .blur(radius: (store.state.isScoop || store.entity.isMine) ? 0 : 12)
-                .cornerRadius(11.16)
-                .padding(EdgeInsets(top: 0, leading: 20.adjusted, bottom: 32.adjustedH, trailing: 20.adjusted))
-        } else {
-            ScrollView(.horizontal, showsIndicators: false) {
-                HStack(spacing: 10.adjusted) {
-                    ForEach(imageList.indices, id: \.self) { index in
-                        RemoteImageView(urlString: imageList[index])
-                            .scaledToFill()
-                            .frame(width: 278.adjusted, height: 278.adjustedH)
-                            .blur(radius: (store.state.isScoop || store.entity.isMine) ? 0 : 12)
-                            .cornerRadius(11.16)
-                    }
-                }
-                .padding(EdgeInsets(top: 0, leading: 20.adjusted, bottom: 32.adjustedH, trailing: 20.adjusted))
-            }
-        }
-    }
-    
-    private var reviewSection: some View {
-        VStack(alignment: .leading, spacing: 8.adjustedH) {
-            
-            IconChip(
-                chip: store.entity.categoryColorResponse.toEntity()
-            )
-            
-            Text("타이틀 없어짐")
-                .customFont(.title2)
-                .foregroundStyle(.black)
-            
-            Text(store.entity.date)
-                .customFont(.caption1m)
-                .foregroundStyle(.gray400)
-            
-            Spacer()
-                .frame(height: 16.adjustedH)
-            
-            Text(
-                (store.state.isScoop || store.entity.isMine)
-                ? store.entity.description.splitZeroWidthSpace()
-                : (store.entity.description.count > 120
-                   ? "\(store.entity.description.prefix(120))...".splitZeroWidthSpace()
-                   : store.entity.description.splitZeroWidthSpace())
-            )
-            .customFont(.body2m)
-            .frame(width: 335.adjusted, alignment: .leading)
-            .foregroundStyle(.black)
-            
-        }
-        .padding(EdgeInsets(top: 0, leading: 20.adjusted, bottom: 32.adjustedH, trailing: 20.adjusted))
-        .frame(maxWidth: .infinity, alignment: .leading)
-    }
-    
-    private var placeInfoSection: some View {
-        VStack(spacing: 0) {
-            menuInfo
-                .background {
-                    Rectangle()
-                        .cornerRadius(20)
-                        .foregroundStyle(.gray0)
-                        .frame(maxHeight: .infinity)
-                }
-                .overlay(alignment: .bottom) {
-                    Line()
-                        .stroke(
-                            style: StrokeStyle(
-                                lineWidth: 1.adjustedH,
-                                dash: [8.adjusted]
-                            )
-                        )
-                        .foregroundStyle(.gray200)
-                        .frame(width: 266.adjusted, height: 1.adjustedH, alignment: .bottom)
-                }
-            
-            locationInfo
-                .background {
-                    Rectangle()
-                        .frame(maxHeight: 134.adjustedH)
-                        .cornerRadius(20)
-                        .foregroundStyle(.gray0)
-                }
-            
-        }
-        .padding(.horizontal, 20.adjusted)
-        .blur(radius: (store.state.isScoop || store.entity.isMine) ? 0 : 12)
-    }
-    
-    private var menuInfo: some View {
-        VStack(alignment: .leading, spacing: 12.adjustedH) {
-            Text("Menu")
-                .customFont(.body1b)
-                .foregroundStyle(.spoonBlack)
-            menuList(menus: store.entity.menuList)
-        }
-        .padding(EdgeInsets(top: 20.adjustedH, leading: 16.adjusted, bottom: 28.adjustedH, trailing: 20.adjusted))
-    }
-    
-    private var locationInfo: some View {
-        HStack {
-            VStack(alignment: .leading, spacing: 12.adjustedH) {
-                Text("Location")
-                    .customFont(.body1b)
-                    .foregroundStyle(.spoonBlack)
-                
-                Text(store.entity.placeName)
-                    .customFont(.title3sb)
-                    .foregroundStyle(.spoonBlack)
-                
-                HStack(spacing: 4.adjusted) {
-                    Image(.icMapGray400)
-                        .resizable()
-                        .scaledToFit()
-                        .frame(width: 20.adjusted, height: 20.adjustedH)
-                    
-                    Text(store.entity.placeAddress)
-                        .customFont(.body2m)
-                        .foregroundStyle(.spoonBlack)
-                }
-            }
-            Spacer()
-        }
-        .padding(EdgeInsets(top: 21.adjustedH, leading: 16.adjusted, bottom: 21.adjustedH, trailing: 16.adjusted))
-    }
-    
-    private var bottomView: some View {
-        HStack(spacing: 0) {
-            SpoonyButton(
-                style: .secondary,
-                size: (store.state.isScoop) ? .medium : .xlarge,
-                title: (store.state.isScoop || store.entity.isMine) ? "길찾기" : "떠먹기",
-                isIcon: (store.state.isScoop || store.entity.isMine) ? false : true,
-                disabled: .constant(false)
-            ) {
-                if store.state.isScoop {
-                    store.send(intent: .pathInfoInNaverMaps)
-                } else {
-//                    navigationManager.popup = .useSpoon(action: {
-//                        store.send(intent: .scoopButtonDidTap)
-//                    })
-                }
-            }
-            
-            if store.state.isScoop {
-                Spacer()
-                
-                ScrapButton(store: store)
-            }
-        }
-        .padding(.horizontal, 20.adjusted)
-    }
-    
-    private var dropDownView: some View {
-        Group {
-            if isPresented {
-                DropDownMenu(
-                    items: ["신고하기"],
-                    isPresented: $isPresented
-                ) { _ in
-                    navigationManager.push(.report(postId: postId))
-                }
-                .frame(alignment: .topTrailing)
-                .padding(.top, 48.adjustedH)
-                .padding(.trailing, 20.adjusted)
-            }
-        }
-    }
-}
-
-struct menuList: View {
-    var menus: [String] = ["고등어봉초밥", "고등어봉초밥", "고등어봉초밥"]
-    
-    var body: some View {
-        VStack(spacing: 12) {
-            ForEach(0..<menus.count, id: \.self) { index in
-                HStack(spacing: 4) {
-                    Image(.icSpoonGray600)
-                        .resizable()
-                        .scaledToFit()
-                        .frame(width: 20.adjusted, height: 20.adjustedH)
-                    Text(menus[index])
-                        .customFont(.body2m)
-                        .lineLimit(2)
-                    Spacer()
-                }
-            }
-        }
-    }
-}
-
-struct ScrapButton: View {
-    
-    @ObservedObject private var store: DetailViewStore
-    
-    init(store: DetailViewStore) {
-        self.store = store
-    }
-    
-    var body: some View {
-        VStack(spacing: 4) {
-            Image(store.state.isZzim ? .icAddmapMain400 : .icAddmapGray400)
-                .resizable()
-                .scaledToFit()
-                .frame(width: 32.adjusted, height: 32.adjustedH)
-                .onTapGesture {
-                    store.send(intent: .scrapButtonDidTap(isScrap: store.state.isZzim))
-                }
-                .padding(EdgeInsets(top: 1.5, leading: 12, bottom: 4, trailing: 12))
-            
-            Text("\(store.state.zzimCount)")
-                .customFont(.caption1m)
-                .foregroundStyle(store.state.isZzim ? .main400 : .gray800)
-                .padding(.bottom, 1.5)
-        }
-    }
-}
-
-struct Line: Shape {
-    func path(in rect: CGRect) -> Path {
-        var path = Path()
-        path.move(to: CGPoint(x: 0, y: 0))
-        path.addLine(to: CGPoint(x: rect.width, y: 0))
-        return path
-    }
-}
-
-#Preview {
-    DetailView(postId: 20)
-        .environmentObject(NavigationManager())
-}
